refactor(tui): split command context into named domain ports (T3)
Split the flat 50-field CommandContext into 10 narrow domain context
types (provider, engine, session, quality, skills, workspace, theme,
agents, permissions, help). Each domain exports a factory that closes
over its context at registration time; Command.run drops its ctx
parameter (now (args, raw) => Promise<void>). createBuiltinCommands is
the composition root that calls every factory.

The dispatcher becomes domain-agnostic: executeCommand takes a
CommandEchoPort (just setMessages for the unknown-command echo) instead
of the full context. The App builds domain context slices separately and
exposes the command list via a Solid signal so the completion controller
reads it reactively without TDZ on the late-initialised list.

Adding a Workspace or Skills command now touches only that domain's
context type and factory, not every command's run signature. Command
bodies are byte-identical (ctx references resolve to the factory
parameter in closure scope); only run(ctx,...) -> run(...) changes.

All aliases, usage, completion contracts, busy behaviors, and output
strings are preserved. Skill/Workspace PTY, contract reactivity, 537
TUI unit tests, and the full baseline (1868 pass / 6 skip / 0 fail)
remain green.
