+++ tests/unit/tui/tui-agent-command.test.ts
@@ -1,12 +1,10 @@
 import { describe, expect, test } from "bun:test";
-import { builtinCommands } from "../../../src/tui/commands/builtin";
-import type { CommandContext } from "../../../src/tui/commands/types";
+import { createAgentsCommands } from "../../../src/tui/commands/agents";
+import type { AgentsCommandContext } from "../../../src/tui/commands/types";
 import type { Message } from "../../../src/tui/types";
 
 describe("/agents command", () => {
   test("renders host Agent Profile and child status output", async () => {
-    const command = builtinCommands.find((entry) => entry.name === "agents");
-    if (!command) throw new Error("Missing /agents command.");
     let messages: Message[] = [];
     const ctx = {
       async agentCommand(args: string) {
@@ -16,9 +14,11 @@ describe("/agents command", () => {
       setMessages(updater: (previous: Message[]) => Message[]) {
         messages = updater(messages);
       },
-    } as unknown as CommandContext;
+    } as unknown as AgentsCommandContext;
+    const command = createAgentsCommands(ctx).find((entry) => entry.name === "agents");
+    if (!command) throw new Error("Missing /agents command.");
 
-    await command.run(ctx, "", "/agents");
+    await command.run("", "/agents");
     expect(messages[0]).toEqual({ role: "user", content: "/agents" });
     expect(messages[1]?.content).toContain("explore-1 [running/background]");
   });
