+++ tests/unit/tui/tui-artifact-command.test.ts
@@ -1,8 +1,11 @@
 import { describe, expect, test } from "bun:test";
 import type { ArtifactEntry } from "../../../src/core/artifacts/workbench";
-import { builtinCommands } from "../../../src/tui/commands/builtin";
-import type { CommandContext } from "../../../src/tui/commands/types";
+import { createBuiltinCommands } from "../../../src/tui/commands/builtin";
+import { createWorkspaceCommands } from "../../../src/tui/commands/workspace";
+import type { BuiltinCommandContexts, WorkspaceCommandContext } from "../../../src/tui/commands/types";
 import type { Message } from "../../../src/tui/types";
+
+const allCommands = createBuiltinCommands({} as unknown as BuiltinCommandContexts);
 
 const entries: ArtifactEntry[] = [
   { path: "workspace/cards/mira.md", updatedAt: "2026-07-10T00:00:00.000Z" },
@@ -11,13 +14,13 @@ const entries: ArtifactEntry[] = [
 
 describe("/artifact command (Workspace page bridge)", () => {
   test("is the only artifact listing command", () => {
-    expect(builtinCommands.some((command) => command.name === "artifact")).toBe(true);
-    expect(builtinCommands.some((command) => command.name === "artifacts")).toBe(false);
+    expect(allCommands.some((command) => command.name === "artifact")).toBe(true);
+    expect(allCommands.some((command) => command.name === "artifacts")).toBe(false);
   });
 
   test("opens the latest artifact in the Workspace page without an argument", async () => {
     const harness = commandHarness();
-    await harness.command.run(harness.ctx, "", "/artifact");
+    await harness.command.run("", "/artifact");
 
     expect(harness.refreshes()).toBe(1);
     expect(harness.openedTargets()).toEqual(["workspace/cards/mira.md"]);
@@ -26,7 +29,7 @@ describe("/artifact command (Workspace page bridge)", () => {
 
   test("reports when there are no artifacts yet", async () => {
     const harness = commandHarness("artifact", []);
-    await harness.command.run(harness.ctx, "", "/artifact");
+    await harness.command.run("", "/artifact");
 
     expect(harness.openedTargets()).toEqual([undefined]);
     expect(harness.messages()[1]?.content).toContain("no artifacts");
@@ -34,7 +37,7 @@ describe("/artifact command (Workspace page bridge)", () => {
 
   test("opens a numbered artifact target in the Workspace page", async () => {
     const harness = commandHarness();
-    await harness.command.run(harness.ctx, "2", "/artifact 2");
+    await harness.command.run("2", "/artifact 2");
 
     expect(harness.openedTargets()).toEqual(["reports/audit.md"]);
     expect(harness.messages()[1]?.content).toContain("reports/audit.md");
@@ -44,7 +47,7 @@ describe("/artifact command (Workspace page bridge)", () => {
     const refreshed = [{ path: "workspace/new.md", updatedAt: "2026-07-21T00:00:00.000Z" }];
     const harness = commandHarness("artifact", refreshed);
 
-    await harness.command.run(harness.ctx, "workspace/new.md", "/artifact workspace/new.md");
+    await harness.command.run("workspace/new.md", "/artifact workspace/new.md");
 
     expect(harness.refreshes()).toBe(1);
     expect(harness.openedTargets()).toEqual(["workspace/new.md"]);
@@ -52,7 +55,7 @@ describe("/artifact command (Workspace page bridge)", () => {
 
   test("keeps a not-found notice in the transcript without switching pages", async () => {
     const harness = commandHarness();
-    await harness.command.run(harness.ctx, "missing.md", "/artifact missing.md");
+    await harness.command.run("missing.md", "/artifact missing.md");
 
     expect(harness.openedTargets()).toEqual([]);
     expect(harness.messages()[1]?.content).toContain("No artifact matches");
@@ -60,8 +63,6 @@ describe("/artifact command (Workspace page bridge)", () => {
 });
 
 function commandHarness(commandName = "artifact", refreshedEntries = entries) {
-  const command = builtinCommands.find((entry) => entry.name === commandName);
-  if (!command) throw new Error("Missing /artifact command.");
   let messages: Message[] = [];
   let refreshCount = 0;
   const openedTargets: (string | undefined)[] = [];
@@ -70,7 +71,6 @@ function commandHarness(commandName = "artifact", refreshedEntries = entries) {
     setMessages(updater: (previous: Message[]) => Message[]) {
       messages = updater(messages);
     },
-    artifacts: () => entries,
     async refreshArtifacts() {
       refreshCount += 1;
       return refreshedEntries;
@@ -79,11 +79,12 @@ function commandHarness(commandName = "artifact", refreshedEntries = entries) {
       openedTargets.push(relPath);
       return relPath ? "file" : null;
     },
-  } as unknown as CommandContext;
+  } as unknown as WorkspaceCommandContext;
+  const command = createWorkspaceCommands(ctx).find((entry) => entry.name === commandName);
+  if (!command) throw new Error(`Missing /${commandName} command.`);
 
   return {
     command,
-    ctx,
     messages: () => messages,
     refreshes: () => refreshCount,
     openedTargets: () => openedTargets,
