+++ tests/unit/tui/tui-permissions-command.test.ts
@@ -1,13 +1,11 @@
 import { describe, expect, test } from "bun:test";
 import type { PermissionMode } from "../../../src/core/permissions";
-import { builtinCommands } from "../../../src/tui/commands/builtin";
-import type { CommandContext } from "../../../src/tui/commands/types";
+import { createPermissionsCommands } from "../../../src/tui/commands/permissions";
+import type { PermissionsCommandContext } from "../../../src/tui/commands/types";
 import type { Message } from "../../../src/tui/types";
 
 describe("/permissions command", () => {
   test("shows and changes the four coarse permission modes", async () => {
-    const command = builtinCommands.find((entry) => entry.name === "permissions");
-    if (!command) throw new Error("Missing /permissions command.");
     let mode: PermissionMode = "MOMENTUM";
     let messages: Message[] = [];
     const ctx = {
@@ -18,18 +16,19 @@ describe("/permissions command", () => {
       async changePermissionMode(next: PermissionMode) {
         mode = next;
       },
-    } as unknown as CommandContext;
+    } as unknown as PermissionsCommandContext;
+    const command = createPermissionsCommands(ctx).find((entry) => entry.name === "permissions");
+    if (!command) throw new Error("Missing /permissions command.");
 
-    await command.run(ctx, "", "/permissions");
+    await command.run("", "/permissions");
     expect(messages.at(-1)?.content).toContain("MOMENTUM");
-    await command.run(ctx, "inertia", "/permissions inertia");
+    await command.run("inertia", "/permissions inertia");
     expect(String(mode)).toBe("INERTIA");
-    await command.run(ctx, "yolo", "/permissions yolo");
+    await command.run("yolo", "/permissions yolo");
     expect(String(mode)).toBe("YOLO");
   });
 
   test("rejects unknown modes before changing state", async () => {
-    const command = builtinCommands.find((entry) => entry.name === "permissions")!;
     let changed = false;
     let messages: Message[] = [];
     const ctx = {
@@ -40,8 +39,9 @@ describe("/permissions command", () => {
       async changePermissionMode() {
         changed = true;
       },
-    } as unknown as CommandContext;
-    await command.run(ctx, "turbo", "/permissions turbo");
+    } as unknown as PermissionsCommandContext;
+    const command = createPermissionsCommands(ctx).find((entry) => entry.name === "permissions")!;
+    await command.run("turbo", "/permissions turbo");
     expect(changed).toBe(false);
     expect(messages.at(-1)?.content).toContain("Unknown permission mode");
   });
