+++ tests/unit/tui/tui-workspace-command.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, test } from "bun:test";
-import { builtinCommands } from "../../../src/tui/commands/builtin";
-import type { CommandContext } from "../../../src/tui/commands/types";
+import { createWorkspaceCommands } from "../../../src/tui/commands/workspace";
+import type { WorkspaceCommandContext } from "../../../src/tui/commands/types";
 import type { Message } from "../../../src/tui/types";
 
 describe("/workspace command", () => {
   test("is registered as immediate so it switches pages while the Agent Loop is busy", () => {
-    const command = builtinCommands.find((entry) => entry.name === "workspace");
+    const command = createWorkspaceCommands({} as unknown as WorkspaceCommandContext).find((entry) => entry.name === "workspace");
     if (!command) throw new Error("Missing /workspace command.");
     const behavior = typeof command.busyBehavior === "function"
       ? command.busyBehavior("")
@@ -14,8 +14,6 @@ describe("/workspace command", () => {
   });
 
   test("opens the Workspace page and echoes the key hints", async () => {
-    const command = builtinCommands.find((entry) => entry.name === "workspace");
-    if (!command) throw new Error("Missing /workspace command.");
     let opened = 0;
     const messages: Message[] = [];
     const ctx = {
@@ -25,9 +23,11 @@ describe("/workspace command", () => {
         messages.push(...updater([]));
         return messages;
       },
-    } as unknown as CommandContext;
+    } as unknown as WorkspaceCommandContext;
+    const command = createWorkspaceCommands(ctx).find((entry) => entry.name === "workspace");
+    if (!command) throw new Error("Missing /workspace command.");
 
-    await command.run(ctx, "", "/workspace");
+    await command.run("", "/workspace");
 
     expect(opened).toBe(1);
     expect(messages[0]).toEqual({ role: "user", content: "/workspace" });
@@ -37,8 +37,6 @@ describe("/workspace command", () => {
   });
 
   test("locates a path argument and reports when it is missing", async () => {
-    const command = builtinCommands.find((entry) => entry.name === "workspace");
-    if (!command) throw new Error("Missing /workspace command.");
     const opened: (string | undefined)[] = [];
     const messages: Message[] = [];
     const ctx = {
@@ -51,14 +49,16 @@ describe("/workspace command", () => {
         messages.push(...updater([]));
         return messages;
       },
-    } as unknown as CommandContext;
+    } as unknown as WorkspaceCommandContext;
+    const command = createWorkspaceCommands(ctx).find((entry) => entry.name === "workspace");
+    if (!command) throw new Error("Missing /workspace command.");
 
-    await command.run(ctx, "workspace/cards", "/workspace workspace/cards");
+    await command.run("workspace/cards", "/workspace workspace/cards");
     expect(opened).toEqual(["workspace/cards"]);
     expect(messages[1]?.content).toContain("Opened workspace/cards");
 
     messages.length = 0;
-    await command.run(ctx, "gone.md", "/workspace gone.md");
+    await command.run("gone.md", "/workspace gone.md");
     expect(messages[1]?.content).toContain("not found");
   });
 });
